printing pointer is working.

+++ ft_printf.h
@@ -3,10 +3,10 @@
 /*                                                        :::      ::::::::   */
 /*   ft_printf.h                                        :+:      :+:    :+:   */
 /*                                                    +:+ +:+         +:+     */
-/*   By: bagovic <bagovic@student.42.fr>            +#+  +:+       +#+        */
+/*   By: berminagovic <berminagovic@student.42.f    +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2021/09/27 09:21:41 by bagovic           #+#    #+#             */
-/*   Updated: 2021/10/31 16:24:25 by bagovic          ###   ########.fr       */
+/*   Updated: 2021/11/02 15:08:21 by berminagovi      ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
@@ -21,7 +21,7 @@ int	g_wildcards[9];
 
 void	ft_initialize_data(void);
 void	ft_putunsigned(long long n);
-int		ft_puthex(long dec, int hexcase);
+int		ft_puthex(unsigned long long dec, int hexcase);
 int		ft_printf(const char *format, ...);
 
 #endif
